+++ apps/test-app/src/routes/briefcase.$fileName.lazy.tsx
@@ -12,14 +12,22 @@ import {
 import { createLazyFileRoute } from "@tanstack/react-router";
 import { useEditorToolSettings } from "../frontend/appui/useEditorToolSettings";
 
-export const Route = createLazyFileRoute("/briefcase")({
-  component: Local,
+export const Route = createLazyFileRoute("/briefcase/$fileName")({
+  component: Briefcase,
 });
 
-function Local() {
+function Briefcase() {
+  const { iModelConnection } = Route.useLoaderData();
   useSyncFrontstageParam();
   useEditorToolSettings();
   const featureOverrides = useFeatureOverrideParams();
+  React.useEffect(() => {
+    const handleBeforeUnload = () => {
+      void iModelConnection.close();
+    };
+    window.addEventListener("beforeunload", handleBeforeUnload);
+    return () => window.removeEventListener("beforeunload", handleBeforeUnload);
+  }, [iModelConnection]);
   return (
     <PageLayout.Content>
       <App featureOverrides={featureOverrides} />