+++ apps/test-app/src/routes/briefcase.$fileName.tsx
@@ -9,14 +9,22 @@ import { appInitializer } from "../frontend/AppInitializer";
 import { UiFramework } from "@itwin/appui-react";
 import { AppParams } from "../frontend/SearchParams";
 import { registerFrontstages } from "../frontend/registerFrontstages";
+import { config } from "../frontend/config";
+import { ProcessDetector } from "@itwin/core-bentley";
 
-export const Route = createFileRoute("/briefcase")({
-  loaderDeps: ({ search: { filePath } }) => ({ filePath }),
+export const Route = createFileRoute("/briefcase/$fileName")({
+  loaderDeps: ({ search }) => {
+    if ("filePath" in search && typeof search.filePath === "string")
+      return { filePath: search.filePath };
+    return {};
+  },
   loader: async (ctx) => {
     await appInitializer.initialize();
 
+    const filePath = ctx.deps.filePath ?? config.bimDir;
+    const fileName = `${filePath}/${ctx.params.fileName}`;
     const iModelConnection = await BriefcaseConnection.openFile({
-      fileName: ctx.deps.filePath,
+      fileName,
     });
     const viewState = await createViewState(iModelConnection);
 
@@ -29,8 +37,18 @@ export const Route = createFileRoute("/briefcase")({
       viewState,
     };
   },
-  validateSearch: (search: AppParams & { filePath: string }) => {
-    return search;
+  onLeave: (ctx) => {
+    const iModelConnection = ctx.loaderData?.iModelConnection;
+    void iModelConnection?.close();
+  },
+  validateSearch: (
+    search: AppParams & { filePath?: string }
+  ): AppParams & { filePath?: string } => {
+    if (ProcessDetector.isElectronAppFrontend) {
+      return search;
+    }
+    const { filePath: _filePath, ...rest } = search;
+    return rest;
   },
   shouldReload: (ctx) => {
     return ctx.cause === "enter";