+++ ui/components-react/src/components-react/new-editors/editors-registry/EditorsRegistryProvider.tsx
@@ -12,8 +12,10 @@ import type { EditorSpec } from "../Types.js";
 import { EditorsRegistryContext } from "./EditorsRegistryContext.js";
 
 /**
- * Provider that adds supplied editors into `EditorsRegistry`. Multiple providers can be nested together.
- * Editors added through the lowest level provider will have the highest priority.
+ * Provider that adds supplied editors into `EditorsRegistry`.
+ * Multiple providers can be nested, with editors from the innermost provider taking the highest priority.
+ * You can either supply a list of editors or provide a function that customizes the current list of editors,
+ * allowing full control over the final set of editors.
  * @beta
  */
 export function EditorsRegistryProvider({
@@ -21,11 +23,14 @@ export function EditorsRegistryProvider({
   editors,
 }: {
   children: React.ReactNode;
-  editors: EditorSpec[];
+  editors: EditorSpec[] | ((editors: EditorSpec[]) => EditorSpec[]);
 }) {
   const parentContext = useContext(EditorsRegistryContext);
 
   const value = useMemo(() => {
+    if (typeof editors === "function") {
+      return { editors: editors(parentContext.editors) };
+    }
     return {
       editors: [...editors, ...parentContext.editors],
     };