+++ ui/imodel-components-react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log - @itwin/imodel-components-react
 
+## 5.7.0
+
+### Patch Changes
+
+- Updated dependencies [fd58c9f]
+  - @itwin/components-react@5.7.0
+  - @itwin/core-react@5.7.0
+
+## 5.6.0
+
+### Patch Changes
+
+- Updated dependencies [a4b74f9]
+  - @itwin/components-react@5.6.0
+  - @itwin/core-react@5.6.0
+
 ## 5.5.0
 
 ### Minor Changes